feat(screenshots): added thumbnail creation for videos

+++ KNARZhelper/FilesCommon/ImageHelper.cs
@@ -1,6 +1,7 @@
 ﻿using ImageMagick;
 using System;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace KNARZhelper.FilesCommon
@@ -10,8 +11,11 @@ namespace KNARZhelper.FilesCommon
     /// </summary>
     internal static class ImageHelper
     {
+        public static readonly string[] SupportedImageExtensions = { ".jpg", ".jpeg", ".png", ".webp", ".bmp" };
+        public static readonly string[] SupportedVideoExtensions = { ".mp4", ".avi", ".webm", ".wmv", ".mov" };
+
         /// <summary>
-        /// Creates a thumbnail image with a height of 120 pixels, maintaining the aspect ratio.
+        /// Creates a thumbnail image.
         /// </summary>
         /// <param name="imageFileName">The path to the original image file.</param>
         /// <param name="thumbNailHeight">Height of the thumbnails that will be generated</param>
@@ -19,11 +23,16 @@ namespace KNARZhelper.FilesCommon
         /// <returns>The FileInfo of the created thumbnail image.</returns>
         public static async Task<FileInfo> CreateThumbnailImage(string imageFileName, int thumbNailHeight, string thumbnailFileName = "")
         {
-            // We exit the method for jxr files and videos, because ImageMagick can't process
-            // themout of the box without serious fiddling or in case of videos at all.
-            if (FileHelper.GetFileExtensionFromUrl(imageFileName).IsOneOf(".jxr", ".mp4", ".avi", ".webm"))
+            //NEXT: Check why the thumbnail file seems to be used. Does a FileInfo lock a file maybe?
+
+            byte[] imageBytes = null;
+            var videoInitialized = false;
+
+            if (SupportedImageExtensions.Contains(FileHelper.GetFileExtensionFromUrl(imageFileName)))
             {
-                return null;
+                imageBytes = imageFileName.IsValidHttpUrl()
+                    ? await FileDownloader.Instance().DownloadFileAsync(new Uri(imageFileName))
+                    : File.ReadAllBytes(imageFileName);
             }
 
             if (string.IsNullOrEmpty(thumbnailFileName))
@@ -34,22 +43,49 @@ namespace KNARZhelper.FilesCommon
 
             var thumbnailFileInfo = new FileInfo(thumbnailFileName);
 
-            if (thumbnailFileInfo.Exists)
+            if (SupportedVideoExtensions.Contains(FileHelper.GetFileExtensionFromUrl(imageFileName)))
             {
-                thumbnailFileInfo.Delete();
+                try
+                {
+                    var ffMpeg = new NReco.VideoConverter.FFMpegConverter();
+
+                    ffMpeg.GetVideoThumbnail(imageFileName, thumbnailFileInfo.FullName, 5);
+
+                    Task.Delay(TimeSpan.FromMilliseconds(100));
+
+                    videoInitialized = true;
+
+                    thumbnailFileInfo.Refresh();
+
+                    if (!thumbnailFileInfo.Exists)
+                    {
+                        return null;
+                    }
+                }
+                catch (Exception ex)
+                {
+                    Log.Error(ex, $"Error processing video {imageFileName}");
+                }
             }
 
-            var imageBytes = imageFileName.IsValidHttpUrl()
-                ? await FileDownloader.Instance().DownloadFileAsync(new Uri(imageFileName))
-                : File.ReadAllBytes(imageFileName);
+            if (!videoInitialized && imageBytes is null)
+            {
+                return null;
+            }
 
             try
             {
-                using (var image = new MagickImage(imageBytes))
+                using (var image = videoInitialized ? new MagickImage(thumbnailFileInfo.FullName) : new MagickImage(imageBytes))
                 {
                     image.Scale(0, (uint)thumbNailHeight);
 
                     image.Format = MagickFormat.Jpg;
+
+                    if (thumbnailFileInfo.Exists)
+                    {
+                        thumbnailFileInfo.Delete();
+                        Task.Delay(TimeSpan.FromMilliseconds(100));
+                    }
 
                     await image.WriteAsync(thumbnailFileName);
                 }
@@ -58,7 +94,7 @@ namespace KNARZhelper.FilesCommon
             }
             catch (Exception ex)
             {
-                Log.Error(ex, $"Error processing file {imageFileName}");
+                Log.Error(ex, $"Error processing image file {imageFileName}");
             }
 
             return null;
